+++ src/gateway/src/main/java/com/mrmccormick/hydra/mqtt/implementation/actor/connector/TagProviderConnector.java
@@ -1,11 +1,14 @@
 package com.mrmccormick.hydra.mqtt.implementation.actor.connector;
 
+import com.google.common.collect.ImmutableList;
 import com.inductiveautomation.ignition.common.BasicDataset;
+import com.inductiveautomation.ignition.common.BundleUtil;
 import com.inductiveautomation.ignition.common.config.*;
 import com.inductiveautomation.ignition.common.document.Document;
 import com.inductiveautomation.ignition.common.i18n.LocalizedString;
 import com.mrmccormick.hydra.mqtt.GatewayHook;
 import com.mrmccormick.hydra.mqtt.domain.Event;
+import com.mrmccormick.hydra.mqtt.domain.EventProperty;
 import com.mrmccormick.hydra.mqtt.domain.actor.ActorBase;
 import com.mrmccormick.hydra.mqtt.domain.actor.connector.IConnector;
 
@@ -38,12 +41,13 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 
-public class TagProviderConnector extends ActorBase implements IConnector, TagChangeListener, TagStructureListener {
+public class TagProviderConnector extends ActorBase implements IConnector, TagChangeListener, TagStructureListener, PropertyModelContributor {
 
     public TagProviderConnector(@NotNull String connectionName,
                                 @NotNull String name,
                                 @NotNull GatewayContext context,
-                                @NotNull String tagProviderName) {
+                                @NotNull String tagProviderName,
+                                boolean publishPropertiesEnabled) {
         super(name);
 
         //noinspection ConstantValue
@@ -60,6 +64,8 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
         if (tagProviderName == null)
             throw new IllegalArgumentException("tagProviderName cannot be null");
         _tagProviderName = tagProviderName;
+
+        _publishPropertiesEnabled = publishPropertiesEnabled;
     }
 
     @Override
@@ -84,6 +90,23 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
         }
 
         return true;
+    }
+
+    @Override
+    public void configure(@Nullable MutableConfigurationPropertyModel mcpm) {
+        if (mcpm == null)
+            return;
+
+        if (_publishPropertiesEnabled) {
+            try {
+                mcpm.addProperties(_propertyRoutingPublishTopicOverride);
+                mcpm.addProperties(_propertyRoutingPublishEnabled);
+                //noinspection unchecked
+                mcpm.registerAllowedValues(_propertyRoutingPublishEnabled, ImmutableList.of(true, false));
+            } catch (Exception e) {
+                _logger.error("Could not configure routing properties: " + e, e);
+            }
+        }
     }
 
     @Override
@@ -125,6 +148,19 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
             return;
         }
 
+        // Load TagEngineConnector.properties values with 'hydramqtt.' prefix
+        BundleUtil.get().addBundle("hydramqtt", this.getClass(), "TagEngineConnector");
+
+        // Register as a class that contributes properties
+        _tagManager.getConfigManager().registerTagPropertyContributor(this);
+
+        // Initialize properties with values from TagEngineConnector.properties
+        initializeProperty(_propertyRoutingPublishEnabled,
+                EventProperty.RoutingPublishEnabled.name(),
+                Boolean.class, true);
+        initializeProperty(_propertyRoutingPublishTopicOverride,
+                EventProperty.RoutingPublishTopicOverride.name(),
+                String.class, "");
         _isConnecting = false;
         _isConnected = true;
     }
@@ -136,6 +172,10 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
             return;
         }
         _isDisconnecting = true;
+
+        BundleUtil.get().removeBundle("hydramqtt");
+
+        _tagManager.getConfigManager().unregisterTagPropertyContributor(this);
 
         try {
             allTagsUnsubscribe();
@@ -255,6 +295,36 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
             return;
         }
 
+        Boolean publishEnabled = null;
+        String publishTopicOverride = null;
+        var tagProperties = tagConfig.getTagProperties();
+        if (_publishPropertiesEnabled) {
+            try {
+                //noinspection unchecked
+                publishEnabled = (Boolean) tagProperties.get(_propertyRoutingPublishEnabled);
+            } catch (Exception e) {
+                _logger.warn("PublishEnabled property returned non-Boolean value: " + e.getMessage(), e);
+            }
+
+            try {
+                //noinspection unchecked
+                publishTopicOverride = (String) tagProperties.get(_propertyRoutingPublishTopicOverride);
+            } catch (Exception e) {
+                _logger.warn("PublishTopic property returned non-String value: " + e.getMessage(), e);
+            }
+        }
+
+        if (publishEnabled != null && !publishEnabled) {
+            return;
+        }
+
+        if (publishEnabled != null)
+            properties.put(EventProperty.RoutingPublishEnabled.name(), publishEnabled);
+        if (publishTopicOverride != null)
+            properties.put(EventProperty.RoutingPublishTopicOverride.name(), publishTopicOverride);
+        properties.put(EventProperty.Documentation.name(), tagProperties.get(WellKnownTagProps.Documentation));
+        properties.put(EventProperty.EngineeringUnits.name(), tagProperties.get(WellKnownTagProps.EngUnit));
+
         var timestamp = tagChangeEvent.getValue().getTimestamp();
         var value = tagChangeEvent.getValue().getValue();
 
@@ -340,6 +410,11 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
     private boolean _isConnecting = false;
     private boolean _isDisconnecting = false;
     private final Logger _logger;
+    private static final BasicConfigurationProperty _propertyRoutingPublishEnabled =
+            new BasicConfigurationProperty<Boolean>();
+    private static final BasicConfigurationProperty _propertyRoutingPublishTopicOverride =
+            new BasicConfigurationProperty<String>();
+    private final Boolean _publishPropertiesEnabled;
     private final GatewayTagManager _tagManager;
     private TagProvider _tagProvider;
     private final String _tagProviderName;
@@ -754,6 +829,16 @@ public class TagProviderConnector extends ActorBase implements IConnector, TagCh
             // Tag has changed type
             needToUpdateTagConfig = true;
             tagConfig.set(WellKnownTagProps.DataType, newTagDataType);
+        }
+
+        if (event.hasProperty(EventProperty.Documentation.name())) {
+            tagConfig.set(WellKnownTagProps.Documentation, (String) event.getPropertyValue(EventProperty.Documentation.name()));
+            needToUpdateTagConfig = true;
+        }
+
+        if (event.hasProperty(EventProperty.EngineeringUnits.name())) {
+            tagConfig.set(WellKnownTagProps.EngUnit, (String) event.getPropertyValue(EventProperty.EngineeringUnits.name()));
+            needToUpdateTagConfig = true;
         }
 
         // If the tag config has changed and needs updated, update it.